Sec. 40 - Deploying Completed
Detail: set environment variable for local and cloud database separately, so the prod version will not be influenced when I change data in local database

need to export in local cmd and set env. var. in heroku with code provided by cloud database

when I use nodemon locally, it will connect to local database, and if I go to heroku address (cloud), it will connect to cloud database (Mongodb Atlas)